requirements.txt
-e . # package from this repo
paho-mqtt
numpy
pytest
